Fix missing icons and subtitle in mobile slideout menu
