+++ app/javascript/mastodon/components/status_action_bar/index.jsx
@@ -24,7 +24,7 @@ import { me } from '../../initial_state';
 
 import { IconButton } from '../icon_button';
 import { isFeatureEnabled } from '../../utils/environment';
-import { ReblogButton } from '../status/reblog_button';
+import { BoostButton } from '../status/reblog_button';
 import { RemoveQuoteHint } from './remove_quote_hint';
 
 const messages = defineMessages({
@@ -372,7 +372,7 @@ class StatusActionBar extends ImmutablePureComponent {
           <IconButton className='status__action-bar__button' title={replyTitle} icon={isReply ? 'reply' : replyIcon} iconComponent={isReply ? ReplyIcon : replyIconComponent} onClick={this.handleReplyClick} counter={status.get('replies_count')} />
         </div>
         <div className='status__action-bar__button-wrapper'>
-          <ReblogButton status={status} counters={withCounters} />
+          <BoostButton status={status} counters={withCounters} />
         </div>
         <div className='status__action-bar__button-wrapper'>
           <IconButton className='status__action-bar__button star-icon' animate active={status.get('favourited')} title={favouriteTitle} icon='star' iconComponent={status.get('favourited') ? StarIcon : StarBorderIcon} onClick={this.handleFavouriteClick} counter={withCounters ? status.get('favourites_count') : undefined} />
