+++ app/javascript/mastodon/models/dropdown_menu.ts
@@ -1,7 +1,13 @@
 import type { KeyboardEvent, MouseEvent, TouchEvent } from 'react';
 
+import type { IconProp } from '../components/icon';
+
 interface BaseMenuItem {
   text: string;
+  description?: string;
+  icon?: IconProp;
+  highlighted?: boolean;
+  disabled?: boolean;
   dangerous?: boolean;
 }
